LayoutTests/imported/w3c/web-platform-tests/css/css-tables/width-distribution/computing-table-width-0-expected.txt

Width Distribution

"Computing the table width"

This is testing that the table root's minimum is max(table-root width, capmin, gridmin) Spec Text


PASS Width is 100px due to 100px content size of caption
